+++ CHANGELOG.md
@@ -19,6 +19,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 > `Security` - in case of vulnerabilities.
 
+## [5.9.5] - 2021-09-29
+
+### Added
+
+- options for initApplicationCommands
+
 ## [5.9.4] - 2021-09-22
 
 ### Added